processing/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sspsygene_processing"
version = "1.0.0"
description = "SSPsyGene Data Processing"
requires-python = ">=3.11"
dependencies = [
    "pandas==2.2.1",
    "pyarrow==21.0.0",
    "click==8.3.0",
    "numpy>=1.26.0,<3.0.0",
    "pyyaml>=6.0.1,<7.0.0",
    "mpmath>=1.3.0,<2.0.0",
    "tqdm>=4.60.0",
    "biopython>=1.83",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
sspsygene = "processing.click.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.package-data]
processing = ["r/*.R"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"
filterwarnings = [
    "ignore::DeprecationWarning:pandas",
]
